# server/src/test/resources/dmo/server/openapi_v1.json
{
  "swagger": "2.0",
  "info": {
    "description": "Api Documentation",
    "version": "1.0",
    "title": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
  "tags": [
    {
      "name": "anime-controller",
      "description": "Anime Controller"
    },
    {
      "name": "anime-status-controller",
      "description": "Anime Status Controller"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/api/v1/anime": {
      "get": {
        "tags": [
          "anime-controller"
        ],
        "summary": "findAll",
        "operationId": "findAllUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "size",
            "in": "query",
            "description": "size",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "title",
            "in": "query",
            "description": "title",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PageDtoOfAnimeLightDto"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/anime/{id}": {
      "get": {
        "tags": [
          "anime-controller"
        ],
        "summary": "getAnime",
        "operationId": "getAnimeUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AnimeDto"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/anime/{id}/episodes": {
      "get": {
        "tags": [
          "anime-controller"
        ],
        "summary": "getEpisodes",
        "operationId": "getEpisodesUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "size",
            "in": "query",
            "description": "size",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PageDtoOfEpisodeDto"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/anime_status": {
      "get": {
        "tags": [
          "anime-status-controller"
        ],
        "summary": "findAll",
        "operationId": "findAllUsingGET_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "accountNonExpired",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "accountNonLocked",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "authorities[0].authority",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "credentialsNonExpired",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "email",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "enabled",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "expiresAfter",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "password",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "size",
            "in": "query",
            "description": "size",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "username",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PageDtoOfAnimeStatusDto"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "anime-status-controller"
        ],
        "summary": "updateStatus",
        "operationId": "updateStatusUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "accountNonExpired",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "accountNonLocked",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "authorities[0].authority",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "credentialsNonExpired",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "email",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "enabled",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "expiresAfter",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "password",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "in": "body",
            "name": "updateAnimeStatusDto",
            "description": "updateAnimeStatusDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateAnimeStatusDto"
            }
          },
          {
            "name": "username",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AnimeStatusDto"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "jwt": {
      "type": "oauth2",
      "tokenUrl": "/login",
      "flow": "password",
      "scopes": {
        "default": "Default auth scope"
      }
    }
  },
  "definitions": {
    "AnimeDto": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "titles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AnimeTitleDto"
          }
        },
        "type": {
          "type": "string",
          "enum": [
            "MOVIE",
            "MUSIC_VIDEO",
            "OTHER",
            "OVA",
            "TV_SERIES",
            "TV_SPECIAL",
            "UNKNOWN",
            "WEB"
          ]
        }
      },
      "title": "AnimeDto"
    },
    "AnimeLightDto": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "titles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AnimeTitleDto"
          }
        },
        "type": {
          "type": "string"
        }
      },
      "title": "AnimeLightDto"
    },
    "AnimeStatusDto": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "status": {
          "type": "string",
          "enum": [
            "COMPLETED",
            "IN_PROGRESS",
            "NOT_STARTED"
          ]
        },
        "titles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AnimeTitleDto"
          }
        },
        "type": {
          "type": "string",
          "enum": [
            "MOVIE",
            "MUSIC_VIDEO",
            "OTHER",
            "OVA",
            "TV_SERIES",
            "TV_SPECIAL",
            "UNKNOWN",
            "WEB"
          ]
        }
      },
      "title": "AnimeStatusDto"
    },
    "AnimeTitleDto": {
      "type": "object",
      "properties": {
        "lang": {
          "type": "string"
        },
        "text": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      },
      "title": "AnimeTitleDto"
    },
    "EpisodeDto": {
      "type": "object",
      "properties": {
        "airDate": {
          "type": "string",
          "format": "date"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "length": {
          "type": "integer",
          "format": "int64"
        },
        "number": {
          "type": "integer",
          "format": "int64"
        },
        "titles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/EpisodeTitleDto"
          }
        },
        "type": {
          "type": "string",
          "enum": [
            "CREDIT",
            "OTHER",
            "PARODY",
            "REGULAR",
            "SPECIAL",
            "TRAILER"
          ]
        }
      },
      "title": "EpisodeDto"
    },
    "EpisodeTitleDto": {
      "type": "object",
      "properties": {
        "lang": {
          "type": "string"
        },
        "text": {
          "type": "string"
        }
      },
      "title": "EpisodeTitleDto"
    },
    "PageDtoOfAnimeLightDto": {
      "type": "object",
      "properties": {
        "content": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AnimeLightDto"
          }
        },
        "number": {
          "type": "integer",
          "format": "int32"
        },
        "numberOfElements": {
          "type": "integer",
          "format": "int32"
        },
        "size": {
          "type": "integer",
          "format": "int32"
        },
        "totalElements": {
          "type": "integer",
          "format": "int64"
        },
        "totalPages": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "PageDtoOfAnimeLightDto"
    },
    "PageDtoOfAnimeStatusDto": {
      "type": "object",
      "properties": {
        "content": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AnimeStatusDto"
          }
        },
        "number": {
          "type": "integer",
          "format": "int32"
        },
        "numberOfElements": {
          "type": "integer",
          "format": "int32"
        },
        "size": {
          "type": "integer",
          "format": "int32"
        },
        "totalElements": {
          "type": "integer",
          "format": "int64"
        },
        "totalPages": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "PageDtoOfAnimeStatusDto"
    },
    "PageDtoOfEpisodeDto": {
      "type": "object",
      "properties": {
        "content": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/EpisodeDto"
          }
        },
        "number": {
          "type": "integer",
          "format": "int32"
        },
        "numberOfElements": {
          "type": "integer",
          "format": "int32"
        },
        "size": {
          "type": "integer",
          "format": "int32"
        },
        "totalElements": {
          "type": "integer",
          "format": "int64"
        },
        "totalPages": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "PageDtoOfEpisodeDto"
    },
    "UpdateAnimeStatusDto": {
      "type": "object",
      "properties": {
        "animeId": {
          "type": "integer",
          "format": "int64"
        },
        "status": {
          "type": "string",
          "enum": [
            "COMPLETED",
            "IN_PROGRESS",
            "NOT_STARTED"
          ]
        }
      },
      "title": "UpdateAnimeStatusDto"
    }
  }
}